# config/kit-service-tinfoil-swagger.json
{
	"swagger": "2.0",
	"info": {
		"description": "Kit Service",
		"version": "1.1.6",
		"title": "Kit Service"
	},
	"host": "kit-service-tinfoil.run.aws-usw02-dev.ice.predix.io",
	"basePath": "/",
	"tags": [{
		"name": "kit-controller",
		"description": "Kit Controller"
	}],
	"schemes": [
		"https"
	],
	"paths": {
		"/device/": {
			"get": {
				"tags": ["kit-controller"],
				"summary": "getDevices",
				"operationId": "getDevicesUsingGET",
				"consumes": ["application/json"],
				"produces": ["*/*"],
				"parameters": [{
					"name": "Authorization",
					"in": "header",
					"description": "Authorization",
					"required": true,
					"type": "string"
				}],
				"responses": {
					"200": {
						"description": "OK",
						"schema": {
							"type": "array",
							"items": {
								"$ref": "#/definitions/RegisterDevice"
							}
						}
					},
					"400": {
						"description": "Bad Request"
					},
					"401": {
						"description": "Unauthorized"
					},
					"403": {
						"description": "Forbidden"
					},
					"404": {
						"description": "Not Found"
					}
				}
			}
		},
		"/device/register": {
			"post": {
				"tags": ["kit-controller"],
				"summary": "registerDevice",
				"operationId": "registerDeviceUsingPOST",
				"consumes": ["application/json"],
				"produces": ["*/*"],
				"parameters": [{
					"in": "body",
					"name": "device",
					"description": "device",
					"required": true,
					"schema": {
						"$ref": "#/definitions/RegisterDevice"
					}
				}, {
					"name": "Authorization",
					"in": "header",
					"description": "Authorization",
					"required": true,
					"type": "string"
				}],
				"responses": {
					"200": {
						"description": "OK",
						"schema": {
							"type": "object"
						}
					},
					"201": {
						"description": "Created"
					},
					"400": {
						"description": "Bad Request"
					},
					"401": {
						"description": "Unauthorized"
					},
					"403": {
						"description": "Forbidden"
					},
					"404": {
						"description": "Not Found"
					}
				}
			}
		},
		"/device/{deviceId}": {
			"get": {
				"tags": ["kit-controller"],
				"summary": "getDevice",
				"operationId": "getDeviceUsingGET",
				"consumes": ["application/json"],
				"produces": ["*/*"],
				"parameters": [{
					"name": "deviceId",
					"in": "path",
					"description": "deviceId",
					"required": true,
					"type": "string"
				}, {
					"name": "Authorization",
					"in": "header",
					"description": "Authorization",
					"required": true,
					"type": "string"
				}],
				"responses": {
					"200": {
						"description": "OK",
						"schema": {
							"type": "object"
						}
					},
					"400": {
						"description": "Bad Request"
					},
					"401": {
						"description": "Unauthorized"
					},
					"403": {
						"description": "Forbidden"
					},
					"404": {
						"description": "Not Found"
					}
				}
			},
			"put": {
				"tags": ["kit-controller"],
				"summary": "updateRegisterDevice",
				"operationId": "updateRegisterDeviceUsingPUT",
				"consumes": ["application/json"],
				"produces": ["*/*"],
				"parameters": [{
					"name": "deviceId",
					"in": "path",
					"description": "deviceId",
					"required": true,
					"type": "string"
				}, {
					"in": "body",
					"name": "device",
					"description": "device",
					"required": true,
					"schema": {
						"$ref": "#/definitions/RegisterDevice"
					}
				}, {
					"name": "Authorization",
					"in": "header",
					"description": "Authorization",
					"required": true,
					"type": "string"
				}],
				"responses": {
					"200": {
						"description": "OK",
						"schema": {
							"type": "object"
						}
					},
					"201": {
						"description": "Created"
					},
					"400": {
						"description": "Bad Request"
					},
					"401": {
						"description": "Unauthorized"
					},
					"403": {
						"description": "Forbidden"
					},
					"404": {
						"description": "Not Found"
					}
				}
			}
		}
	},
	"definitions": {
		"AssetTag": {
			"type": "object",
			"properties": {
				"alertStatusUri": {
					"type": "string"
				},
				"edgeDatasource": {
					"$ref": "#/definitions/EdgeDatasource"
				},
				"hiAlarmThreshold": {
					"type": "number",
					"format": "double"
				},
				"hiQualityThreshold": {
					"type": "number",
					"format": "double"
				},
				"isKpi": {
					"type": "string"
				},
				"label": {
					"type": "string"
				},
				"lastCalibrated": {
					"$ref": "#/definitions/XMLGregorianCalendar"
				},
				"loAlarmThreshold": {
					"type": "number",
					"format": "double"
				},
				"loQualityThreshold": {
					"type": "number",
					"format": "double"
				},
				"locationUUID": {
					"type": "string"
				},
				"tagUri": {
					"type": "string"
				},
				"timeseriesDatasource": {
					"$ref": "#/definitions/TimeseriesDatasource"
				},
				"unit": {
					"type": "string"
				}
			}
		},
		"XMLGregorianCalendar": {
			"type": "object",
			"properties": {
				"day": {
					"type": "integer",
					"format": "int32"
				},
				"fractionalSecond": {
					"type": "number",
					"format": "double"
				},
				"hour": {
					"type": "integer",
					"format": "int32"
				},
				"millisecond": {
					"type": "integer",
					"format": "int32"
				},
				"minute": {
					"type": "integer",
					"format": "int32"
				},
				"month": {
					"type": "integer",
					"format": "int32"
				},
				"second": {
					"type": "integer",
					"format": "int32"
				},
				"timezone": {
					"type": "integer",
					"format": "int32"
				}
			}
		},
		"TimeseriesDatasource": {
			"type": "object",
			"properties": {
				"tag": {
					"type": "string"
				}
			}
		},
		"RegisterDevice": {
			"type": "object",
			"properties": {
				"activationDate": {
					"type": "string"
				},
				"createdDate": {
					"type": "string"
				},
				"deviceAddress": {
					"type": "string"
				},
				"deviceConfig": {
					"type": "object"
				},
				"deviceName": {
					"type": "string"
				},
				"deviceType": {
					"type": "string"
				},
				"groupRef": {
					"type": "string"
				},
				"tags": {
					"type": "array",
					"items": {
						"$ref": "#/definitions/AssetTag"
					}
				},
				"updateDate": {
					"type": "string"
				},
				"uri": {
					"type": "string"
				},
				"userId": {
					"type": "string"
				}
			}
		},
		"EdgeDatasource": {
			"type": "object",
			"properties": {
				"controllerUri": {
					"type": "string"
				},
				"nodeName": {
					"type": "string"
				}
			}
		}
	}
}